diet(phase6): split dispatch/response_support.rs (1409 LOC) into 6-module subdir
Mechanical split, behavior unchanged. response.rs use-statement unchanged
(mod.rs re-exports the same 10 symbols).

dispatch/response_support/
- mod.rs (14)              re-exports only
- budget.rs (49)           effective_budget_for_tool, budget_hint,
                           max_result_size_chars_for_tool
- routing_hint.rs (67)     apply_contextual_guidance, routing_hint_for_payload
- payload_compact.rs (319) strip_empty_fields, compact_response_payload,
                           summarize_* helpers
- text_channel.rs (737)    text_payload_for_response, slim_text_payload_*,
                           format_structured_response + text_channel_tests
- truncation.rs (238)      bounded_result_payload, TruncationInfo,
                           enrich_recovery_hint_for_signals,
                           recovery_hint_for_stage + routing_hint_tier_tests
- envelope.rs (25)         success_jsonrpc_response

Gates passed locally:
- cargo check --workspace --features http
- cargo test -p codelens-mcp --bin codelens-mcp (593 passed, 0 failed)
- cargo fmt --all -- --check
