+++ crates/codelens-mcp/src/dispatch/response_support/routing_hint.rs
@@ -1,0 +1,67 @@
+use crate::protocol::{RoutingHint, ToolCallResponse};
+use crate::tool_defs::ToolSurface;
+use crate::tools;
+
+pub(crate) fn apply_contextual_guidance(
+    resp: &mut ToolCallResponse,
+    name: &str,
+    recent_tools: &[String],
+    harness_phase: Option<&str>,
+    surface: ToolSurface,
+) -> bool {
+    resp.suggested_next_tools = tools::suggest_next_contextual(name, recent_tools, harness_phase);
+
+    let mut emitted_composite_guidance = false;
+    if let Some((guided_tools, guidance_hint)) =
+        tools::composite_guidance_for_chain(name, recent_tools, surface)
+    {
+        emitted_composite_guidance = true;
+        let mut suggestions = guided_tools;
+        if let Some(existing) = resp.suggested_next_tools.take() {
+            for tool in existing {
+                if suggestions.len() >= 3 {
+                    break;
+                }
+                if !suggestions.iter().any(|candidate| candidate == &tool) {
+                    suggestions.push(tool);
+                }
+            }
+        }
+        resp.suggested_next_tools = Some(suggestions);
+        resp.budget_hint = Some(match resp.budget_hint.take() {
+            Some(existing) => format!("{existing} {guidance_hint}"),
+            None => guidance_hint,
+        });
+    }
+    emitted_composite_guidance
+}
+
+pub(crate) fn routing_hint_for_payload(resp: &ToolCallResponse) -> RoutingHint {
+    let data = resp.data.as_ref();
+    let is_cached = data
+        .and_then(|d| d.get("reused"))
+        .and_then(|v| v.as_bool())
+        .unwrap_or(false);
+    let cache_tier = data
+        .and_then(|d| d.get("cache_hit_tier"))
+        .and_then(|v| v.as_str());
+    let is_async_job = data
+        .and_then(|d| d.get("job_id"))
+        .and_then(|v| v.as_str())
+        .is_some();
+    let is_analysis_handle = data
+        .and_then(|d| d.get("analysis_id"))
+        .and_then(|v| v.as_str())
+        .is_some();
+    if is_cached {
+        match cache_tier {
+            Some("exact") => RoutingHint::CachedExact,
+            Some("warm") => RoutingHint::CachedWarm,
+            _ => RoutingHint::Cached,
+        }
+    } else if is_async_job || is_analysis_handle {
+        RoutingHint::Async
+    } else {
+        RoutingHint::Sync
+    }
+}
