+++ crates/codelens-mcp/src/dispatch/response_support/text_channel.rs
@@ -1,160 +1,20 @@
-use crate::protocol::{JsonRpcResponse, RoutingHint, ToolCallResponse};
-use crate::tool_defs::{ToolSurface, tool_definition};
-use crate::tools;
-use serde_json::{Map, Value, json};
+use crate::protocol::ToolCallResponse;
+use serde_json::{Map, Value};
 
-pub(crate) fn effective_budget_for_tool(name: &str, request_budget: usize) -> usize {
-    tool_definition(name)
-        .and_then(|t| t.max_response_tokens)
-        .map(|cap| request_budget.min(cap))
-        .unwrap_or(request_budget)
-}
+use super::payload_compact::{
+    copy_summarized_field, insert_if_present, summarize_text_data_for_response,
+};
 
-pub(crate) fn budget_hint(tool_name: &str, tokens: usize, budget: usize) -> String {
-    if matches!(
-        tool_name,
-        "get_project_structure" | "get_symbols_overview" | "get_current_config" | "onboard_project"
-    ) {
-        return "overview complete — drill into specific files or symbols".to_owned();
-    }
-    let pct = tokens
-        .checked_mul(100)
-        .and_then(|v| v.checked_div(budget))
-        .unwrap_or(100);
-    let base = format!("{tokens} tokens ({pct}% of {budget} budget)");
+#[cfg(test)]
+use super::payload_compact::{
+    TEXT_CHANNEL_MAX_ARRAY_ITEMS, summarize_structured_content, summarize_text_object,
+};
 
-    if pct > 95 {
-        format!(
-            "{base}. Response near limit — use get_analysis_section to expand specific parts instead of full reports."
-        )
-    } else if pct > 75 {
-        format!("{base}. Consider narrowing scope with path or max_tokens parameter.")
-    } else {
-        base
-    }
-}
-
-pub(crate) fn apply_contextual_guidance(
-    resp: &mut ToolCallResponse,
-    name: &str,
-    recent_tools: &[String],
-    harness_phase: Option<&str>,
-    surface: ToolSurface,
-) -> bool {
-    resp.suggested_next_tools = tools::suggest_next_contextual(name, recent_tools, harness_phase);
-
-    let mut emitted_composite_guidance = false;
-    if let Some((guided_tools, guidance_hint)) =
-        tools::composite_guidance_for_chain(name, recent_tools, surface)
-    {
-        emitted_composite_guidance = true;
-        let mut suggestions = guided_tools;
-        if let Some(existing) = resp.suggested_next_tools.take() {
-            for tool in existing {
-                if suggestions.len() >= 3 {
-                    break;
-                }
-                if !suggestions.iter().any(|candidate| candidate == &tool) {
-                    suggestions.push(tool);
-                }
-            }
-        }
-        resp.suggested_next_tools = Some(suggestions);
-        resp.budget_hint = Some(match resp.budget_hint.take() {
-            Some(existing) => format!("{existing} {guidance_hint}"),
-            None => guidance_hint,
-        });
-    }
-    emitted_composite_guidance
-}
-
-pub(crate) fn routing_hint_for_payload(resp: &ToolCallResponse) -> RoutingHint {
-    let data = resp.data.as_ref();
-    let is_cached = data
-        .and_then(|d| d.get("reused"))
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-    let cache_tier = data
-        .and_then(|d| d.get("cache_hit_tier"))
-        .and_then(|v| v.as_str());
-    let is_async_job = data
-        .and_then(|d| d.get("job_id"))
-        .and_then(|v| v.as_str())
-        .is_some();
-    let is_analysis_handle = data
-        .and_then(|d| d.get("analysis_id"))
-        .and_then(|v| v.as_str())
-        .is_some();
-    if is_cached {
-        match cache_tier {
-            Some("exact") => RoutingHint::CachedExact,
-            Some("warm") => RoutingHint::CachedWarm,
-            _ => RoutingHint::Cached,
-        }
-    } else if is_async_job || is_analysis_handle {
-        RoutingHint::Async
-    } else {
-        RoutingHint::Sync
-    }
-}
-
-/// Recursively strip empty arrays, null values, and empty strings from a JSON Value.
-pub(crate) fn strip_empty_fields(value: &mut serde_json::Value) {
-    match value {
-        serde_json::Value::Object(map) => {
-            map.retain(|_, v| {
-                strip_empty_fields(v);
-                !is_empty_value(v)
-            });
-        }
-        serde_json::Value::Array(arr) => {
-            for item in arr.iter_mut() {
-                strip_empty_fields(item);
-            }
-        }
-        _ => {}
-    }
-}
-
-fn is_empty_value(v: &serde_json::Value) -> bool {
-    match v {
-        serde_json::Value::Null => true,
-        serde_json::Value::String(s) => s.is_empty(),
-        serde_json::Value::Array(a) => a.is_empty(),
-        serde_json::Value::Object(m) => m.is_empty(),
-        _ => false,
-    }
-}
-
-pub(crate) fn compact_response_payload(resp: &mut ToolCallResponse) {
-    if let Some(ref mut data) = resp.data {
-        strip_empty_fields(data);
-    }
-    if let Some(ref mut data) = resp.data
-        && let Some(obj) = data.as_object_mut()
-    {
-        obj.remove("quality_focus");
-        obj.remove("recommended_checks");
-        obj.remove("performance_watchpoints");
-        obj.remove("available_sections");
-        obj.remove("evidence_handles");
-        obj.remove("schema_version");
-        obj.remove("report_kind");
-        obj.remove("profile");
-        obj.remove("next_actions");
-        obj.remove("machine_summary");
-        if let Some(checks) = obj.get_mut("verifier_checks")
-            && let Some(arr) = checks.as_array_mut()
-        {
-            for check in arr.iter_mut() {
-                if let Some(check_obj) = check.as_object_mut() {
-                    check_obj.remove("summary");
-                    check_obj.remove("evidence_section");
-                }
-            }
-        }
-    }
-}
+#[cfg(test)]
+use super::truncation::{
+    TruncationInfo, bounded_result_payload, enrich_recovery_hint_for_signals,
+    recovery_hint_for_stage,
+};
 
 pub(crate) fn text_payload_for_response(
     resp: &ToolCallResponse,
@@ -477,474 +337,6 @@ fn slim_text_payload_for_async_handle(
     }
 
     Some(Value::Object(payload))
-}
-
-fn insert_if_present(target: &mut Map<String, Value>, key: &str, value: Option<Value>) {
-    if let Some(value) = value {
-        target.insert(key.to_owned(), value);
-    }
-}
-
-fn copy_summarized_field(target: &mut Map<String, Value>, source: &Map<String, Value>, key: &str) {
-    if let Some(value) = source.get(key) {
-        target.insert(key.to_owned(), summarize_structured_content(value, 0));
-    }
-}
-
-fn copy_raw_field(target: &mut Map<String, Value>, source: &Map<String, Value>, key: &str) {
-    if let Some(value) = source.get(key) {
-        target.insert(key.to_owned(), value.clone());
-    }
-}
-
-fn summarize_text_data_for_response(value: &Value) -> Value {
-    let Some(object) = value.as_object() else {
-        return summarize_structured_content(value, 0);
-    };
-
-    if object.contains_key("capabilities")
-        && object.contains_key("visible_tools")
-        && object.contains_key("routing")
-    {
-        return summarize_bootstrap_text_data(object);
-    }
-
-    summarize_text_object(object, 0)
-}
-
-fn summarize_bootstrap_text_data(source: &Map<String, Value>) -> Value {
-    let mut out = Map::new();
-    copy_summarized_field(&mut out, source, "active_surface");
-    copy_summarized_field(&mut out, source, "token_budget");
-    copy_summarized_field(&mut out, source, "index_recovery");
-    copy_summarized_field(&mut out, source, "health_summary");
-    copy_summarized_field(&mut out, source, "warnings");
-
-    if let Some(project) = source.get("project").and_then(|value| value.as_object()) {
-        let mut project_summary = Map::new();
-        copy_summarized_field(&mut project_summary, project, "project_name");
-        copy_summarized_field(&mut project_summary, project, "backend_id");
-        copy_summarized_field(&mut project_summary, project, "indexed_files");
-        if !project_summary.is_empty() {
-            out.insert("project".to_owned(), Value::Object(project_summary));
-        }
-    }
-
-    if let Some(capabilities) = source
-        .get("capabilities")
-        .and_then(|value| value.as_object())
-    {
-        let mut capability_summary = Map::new();
-        copy_summarized_field(&mut capability_summary, capabilities, "index_fresh");
-        copy_summarized_field(&mut capability_summary, capabilities, "indexed_files");
-        copy_summarized_field(&mut capability_summary, capabilities, "supported_files");
-        copy_summarized_field(&mut capability_summary, capabilities, "stale_files");
-        copy_summarized_field(
-            &mut capability_summary,
-            capabilities,
-            "semantic_search_status",
-        );
-        copy_summarized_field(&mut capability_summary, capabilities, "lsp_attached");
-        copy_summarized_field(&mut capability_summary, capabilities, "health_summary");
-        if !capability_summary.is_empty() {
-            out.insert("capabilities".to_owned(), Value::Object(capability_summary));
-        }
-    }
-
-    if let Some(visible_tools) = source
-        .get("visible_tools")
-        .and_then(|value| value.as_object())
-    {
-        let mut tools_summary = Map::new();
-        copy_summarized_field(&mut tools_summary, visible_tools, "tool_count");
-        copy_summarized_field(&mut tools_summary, visible_tools, "tool_names");
-        copy_summarized_field(
-            &mut tools_summary,
-            visible_tools,
-            "tool_names_omitted_count",
-        );
-        copy_summarized_field(&mut tools_summary, visible_tools, "effective_namespaces");
-        copy_summarized_field(&mut tools_summary, visible_tools, "effective_tiers");
-        if !tools_summary.is_empty() {
-            out.insert("visible_tools".to_owned(), Value::Object(tools_summary));
-        }
-    }
-
-    if let Some(routing) = source.get("routing").and_then(|value| value.as_object()) {
-        let mut routing_summary = Map::new();
-        copy_summarized_field(&mut routing_summary, routing, "recommended_entrypoint");
-        copy_summarized_field(&mut routing_summary, routing, "preferred_entrypoints");
-        copy_summarized_field(
-            &mut routing_summary,
-            routing,
-            "preferred_entrypoints_visible",
-        );
-        copy_summarized_field(
-            &mut routing_summary,
-            routing,
-            "preferred_entrypoints_visible_omitted_count",
-        );
-        // Omission records are already compact, machine-readable recovery
-        // hints (`reason`, `recommended_action`, surface/profile). Preserve
-        // them verbatim for text-only MCP clients so the fallback channel
-        // remains operationally equivalent to `structuredContent`.
-        copy_raw_field(
-            &mut routing_summary,
-            routing,
-            "preferred_entrypoints_omitted",
-        );
-        if !routing_summary.is_empty() {
-            out.insert("routing".to_owned(), Value::Object(routing_summary));
-        }
-    }
-
-    Value::Object(out)
-}
-
-fn summarize_text_object(source: &Map<String, Value>, depth: usize) -> Value {
-    const MAX_OBJECT_ITEMS: usize = 8;
-
-    let mut summarized = Map::new();
-    let mut array_shrunk = false;
-    // #211: previously the loop broke on the first cap hit and only
-    // marked `truncated: true`, leaving callers with no way to know
-    // which keys had been dropped. Collect every dropped key so the
-    // response carries a `_omitted_keys` list and downstream agents
-    // can request the missing fields explicitly.
-    let mut omitted_keys: Vec<String> = Vec::new();
-    for (index, (key, value)) in source.iter().enumerate() {
-        if index >= MAX_OBJECT_ITEMS {
-            omitted_keys.push(key.clone());
-            continue;
-        }
-        if let Value::Array(items) = value
-            && items.len() > TEXT_CHANNEL_MAX_ARRAY_ITEMS
-        {
-            array_shrunk = true;
-        }
-        summarized.insert(key.clone(), summarize_text_value(value, depth + 1));
-    }
-    if !omitted_keys.is_empty() {
-        summarized.insert("truncated".to_owned(), Value::Bool(true));
-        summarized.insert("_omitted_keys".to_owned(), json!(omitted_keys));
-    }
-    if source.contains_key("truncated") || array_shrunk {
-        summarized.insert("truncated".to_owned(), Value::Bool(true));
-    }
-    Value::Object(summarized)
-}
-
-fn summarize_text_value(value: &Value, depth: usize) -> Value {
-    match value {
-        Value::Object(map) => summarize_text_object(map, depth),
-        other => summarize_structured_content(other, depth),
-    }
-}
-
-/// Adaptive compression based on OpenDev 5-stage strategy (arxiv:2603.05344).
-/// Thresholds are adjusted by effort level offset (Low=-10, Medium=0, High=+10).
-/// Stage 1: <75% budget → pass through
-/// Stage 2: 75-85% → summarize structured content (depth=1)
-/// Stage 3: 85-95% → aggressive summarize (depth=0)
-/// Stage 4: 95-100% → drop structured content entirely
-/// Stage 5: >100% → hard truncation to error payload
-///
-/// Returns (text, structured_content, truncation_info). When the payload
-/// passes through (stage 1), `truncation_info` is `None`. When any
-/// summarization or truncation happens, `truncation_info` carries the
-/// stage, original payload size estimate, effective budget, and a
-/// human-readable recovery hint so callers can surface the loss to
-/// agents at the top level instead of burying `truncated: true` in the
-/// data envelope.
-pub(crate) fn bounded_result_payload(
-    mut text: String,
-    mut structured_content: Option<Value>,
-    payload_estimate: usize,
-    effective_budget: usize,
-    effort_offset: i32,
-) -> (String, Option<Value>, Option<TruncationInfo>) {
-    let usage_pct = payload_estimate
-        .checked_mul(100)
-        .and_then(|v| v.checked_div(effective_budget))
-        .unwrap_or(100);
-    // Apply effort offset to thresholds (High effort delays compression)
-    let t1 = (75i32 + effort_offset).clamp(50, 90) as usize;
-    let t2 = (85i32 + effort_offset).clamp(60, 95) as usize;
-    let t3 = (95i32 + effort_offset).clamp(70, 100) as usize;
-    let t4 = (100i32 + effort_offset).clamp(80, 110) as usize;
-
-    let max_chars = effective_budget * 8;
-    let stage: u8;
-
-    if usage_pct <= t1 {
-        // Stage 1: pass through
-        stage = 1;
-    } else if usage_pct <= t2 {
-        // Stage 2: light summarization
-        stage = 2;
-        if let Some(existing) = structured_content.as_ref() {
-            structured_content = Some(summarize_structured_content(existing, 1));
-        }
-    } else if usage_pct <= t3 {
-        // Stage 3: aggressive summarization
-        stage = 3;
-        if let Some(existing) = structured_content.as_ref() {
-            structured_content = Some(summarize_structured_content(existing, 0));
-        }
-    } else if usage_pct <= t4 {
-        // Stage 4: aggressive summarize structured + truncate text if needed
-        stage = 4;
-        if let Some(existing) = structured_content.as_ref() {
-            structured_content = Some(summarize_structured_content(existing, 0));
-        }
-        if text.len() > max_chars {
-            let byte_idx = text
-                .char_indices()
-                .nth(max_chars)
-                .map(|(i, _)| i)
-                .unwrap_or(text.len());
-            text.truncate(byte_idx);
-            text.push_str("...[truncated]");
-        }
-    } else {
-        // Stage 5: hard truncation — summarize structured to minimal skeleton
-        stage = 5;
-        if let Some(existing) = structured_content.as_ref() {
-            structured_content = Some(summarize_structured_content(existing, 0));
-        }
-        text = serde_json::to_string(&json!({
-            "success": true,
-            "truncated": true,
-            "compression_stage": 5,
-            "error": format!(
-                "Response too large ({} tokens, budget {}). Narrow with path, max_tokens, or depth.",
-                payload_estimate, effective_budget
-            ),
-            "token_estimate": payload_estimate,
-        }))
-        .unwrap_or_else(|_| "{\"success\":false,\"truncated\":true}".to_owned());
-    }
-
-    let info = if stage >= 2 {
-        Some(TruncationInfo {
-            stage,
-            original_payload_estimate: payload_estimate,
-            effective_budget,
-            recovery_hint: recovery_hint_for_stage(stage, payload_estimate, effective_budget),
-        })
-    } else {
-        None
-    };
-    (text, structured_content, info)
-}
-
-/// Top-level metadata describing what the adaptive compressor did.
-///
-/// This is surfaced into the structured response so that an agent can
-/// detect "I asked for X but only got a summarized view" without having
-/// to reach into the data envelope. Stage 1 (pass-through) returns
-/// `None` from `bounded_result_payload`; stages 2–5 emit one of these.
-#[derive(Debug, Clone)]
-pub(crate) struct TruncationInfo {
-    pub stage: u8,
-    pub original_payload_estimate: usize,
-    pub effective_budget: usize,
-    pub recovery_hint: String,
-}
-
-impl TruncationInfo {
-    pub fn to_json(&self) -> Value {
-        json!({
-            "compression_stage": self.stage,
-            "original_payload_estimate": self.original_payload_estimate,
-            "effective_budget": self.effective_budget,
-            "recovery_hint": self.recovery_hint,
-        })
-    }
-}
-
-/// When the call-graph extractor produced only `unresolved` edges and the
-/// response was clipped, the default `recovery_hint` ("narrow with file_path
-/// or smaller max_results") is misleading — the issue is extractor coverage,
-/// not budget. Append a tool-aware grep cue that names the actual symbol so
-/// an agent can fall back to direct text search instead of retrying the same
-/// low-confidence call-graph query with smaller bounds.
-///
-/// No-op for stages < 4, missing structured content, or non-`unresolved_only`
-/// confidence bases — the existing hint is correct in those cases.
-pub(crate) fn enrich_recovery_hint_for_signals(
-    info: TruncationInfo,
-    structured_content: Option<&Value>,
-) -> TruncationInfo {
-    if info.stage < 4 {
-        return info;
-    }
-    let Some(content) = structured_content.and_then(|v| v.as_object()) else {
-        return info;
-    };
-    let basis = content
-        .get("confidence_basis")
-        .and_then(Value::as_str)
-        .unwrap_or("");
-    if basis != "unresolved_only" {
-        return info;
-    }
-    let symbol = content
-        .get("function")
-        .or_else(|| content.get("symbol_name"))
-        .or_else(|| content.get("name"))
-        .and_then(Value::as_str)
-        .unwrap_or("the symbol");
-    let mut recovery_hint = info.recovery_hint;
-    recovery_hint.push_str(&format!(
-        " Call graph returned only `unresolved` edges and the result was clipped — `Grep '{symbol}'` directly may surface raw matches the tree-sitter call query missed."
-    ));
-    TruncationInfo {
-        recovery_hint,
-        ..info
-    }
-}
-
-fn recovery_hint_for_stage(stage: u8, estimate: usize, budget: usize) -> String {
-    match stage {
-        2 => format!(
-            "Light summarization applied ({} of {} budget). Drill into a specific file or symbol for full detail.",
-            estimate, budget
-        ),
-        3 => format!(
-            "Aggressive summarization applied ({} of {} budget). Arrays clipped to 3 items each — use file_path / max_results to narrow scope.",
-            estimate, budget
-        ),
-        4 => format!(
-            "Near-budget summarization applied ({} of {} budget). Result arrays clipped — narrow scope with file_path or smaller max_results to recover the full set.",
-            estimate, budget
-        ),
-        5 => format!(
-            "Response over budget ({} tokens vs {}); structured arrays clipped to 3 items each. Use file_path / smaller max_results / get_analysis_section to recover items.",
-            estimate, budget
-        ),
-        _ => "Compression applied".to_owned(),
-    }
-}
-
-/// Determine `_meta["anthropic/maxResultSizeChars"]` based on tool tier.
-/// Claude Code v2.1.91+ respects this annotation to keep up to 500K chars.
-pub(crate) fn max_result_size_chars_for_tool(name: &str, truncated: bool) -> usize {
-    use crate::protocol::ToolTier;
-    use crate::tool_defs::tool_tier;
-
-    if truncated {
-        return 25_000;
-    }
-
-    match tool_tier(name) {
-        ToolTier::Workflow => 200_000,
-        ToolTier::Analysis => 100_000,
-        ToolTier::Primitive => 50_000,
-    }
-}
-
-pub(crate) fn success_jsonrpc_response(
-    id: Option<Value>,
-    tool_name: &str,
-    text: String,
-    structured_content: Option<Value>,
-    max_result_size_chars: Option<usize>,
-) -> JsonRpcResponse {
-    let mut result = json!({
-        "content": [{ "type": "text", "text": text }]
-    });
-    if let Some(structured_content) = structured_content {
-        result["structuredContent"] = structured_content;
-    }
-    result["_meta"] = json!({
-        "codelens/preferredExecutor": crate::tool_defs::tool_preferred_executor_label(tool_name)
-    });
-    if let Some(max_chars) = max_result_size_chars {
-        result["_meta"]["anthropic/maxResultSizeChars"] = json!(max_chars);
-    }
-    crate::tool_defs::apply_tool_deprecation_meta(&mut result["_meta"], tool_name);
-    JsonRpcResponse::result(id, result)
-}
-
-// Keeping `count`/`returned_count` metadata consistent with the visible array
-// requires `summarize_text_object` and `summarize_structured_content` to share
-// the same cap, so the shrink detector and the shrinker never disagree.
-const TEXT_CHANNEL_MAX_ARRAY_ITEMS: usize = 3;
-
-fn summarize_structured_content(value: &Value, depth: usize) -> Value {
-    const MAX_STRING_CHARS: usize = 240;
-    const MAX_ARRAY_ITEMS: usize = TEXT_CHANNEL_MAX_ARRAY_ITEMS;
-    const MAX_OBJECT_DEPTH: usize = 4;
-
-    match value {
-        Value::Null | Value::Bool(_) | Value::Number(_) => value.clone(),
-        Value::String(text) => {
-            if text.chars().count() <= MAX_STRING_CHARS {
-                value.clone()
-            } else {
-                let truncated = text.chars().take(MAX_STRING_CHARS).collect::<String>();
-                Value::String(format!("{truncated}..."))
-            }
-        }
-        Value::Array(items) => Value::Array(
-            items
-                .iter()
-                .take(MAX_ARRAY_ITEMS)
-                .map(|item| summarize_structured_content(item, depth + 1))
-                .collect(),
-        ),
-        Value::Object(map) => {
-            let max_items = if depth >= MAX_OBJECT_DEPTH {
-                MAX_ARRAY_ITEMS
-            } else {
-                usize::MAX
-            };
-            let mut summarized = serde_json::Map::with_capacity(map.len().min(max_items));
-            // Track sibling `<key>_omitted_count` markers so an agent
-            // sees both the top-level `truncation_warning` AND the
-            // per-array original size at the call site (e.g. callers
-            // array clipped from 287 → 3 records `callers_omitted_count: 284`).
-            let mut omitted_markers: Vec<(String, usize)> = Vec::new();
-            for (index, (key, item)) in map.iter().enumerate() {
-                if index >= max_items {
-                    break;
-                }
-                if should_preserve_structured_array(key, item) {
-                    summarized.insert(key.clone(), item.clone());
-                    continue;
-                }
-                if let Value::Array(items) = item
-                    && items.len() > MAX_ARRAY_ITEMS
-                {
-                    omitted_markers.push((
-                        format!("{key}_omitted_count"),
-                        items.len() - MAX_ARRAY_ITEMS,
-                    ));
-                }
-                summarized.insert(key.clone(), summarize_structured_content(item, depth + 1));
-            }
-            for (marker_key, omitted) in omitted_markers {
-                if !summarized.contains_key(&marker_key) {
-                    summarized.insert(marker_key, json!(omitted));
-                }
-            }
-            if map.contains_key("truncated") {
-                summarized.insert("truncated".to_owned(), Value::Bool(true));
-            }
-            Value::Object(summarized)
-        }
-    }
-}
-
-fn should_preserve_structured_array(key: &str, value: &Value) -> bool {
-    matches!(
-        key,
-        "preferred_entrypoints"
-            | "preferred_entrypoints_visible"
-            | "preferred_entrypoints_omitted"
-            | "preferred_entrypoints_with_executors"
-    ) && value.is_array()
 }
 
 #[cfg(test)]
@@ -1341,69 +733,5 @@ mod text_channel_tests {
         let summarized = summarize_structured_content(&payload, 0);
         let obj = summarized.as_object().expect("object");
         assert!(obj.get("callers_omitted_count").is_none());
-    }
-}
-
-#[cfg(test)]
-mod routing_hint_tier_tests {
-    use super::routing_hint_for_payload;
-    use crate::protocol::{
-        AnalysisSource, BackendKind, Freshness, RoutingHint, ToolCallResponse, ToolResponseMeta,
-    };
-    use serde_json::json;
-
-    fn resp_with_data(data: serde_json::Value) -> ToolCallResponse {
-        let meta = ToolResponseMeta {
-            backend_used: BackendKind::Hybrid.to_string(),
-            confidence: 1.0,
-            degraded_reason: None,
-            source: AnalysisSource::Native,
-            partial: false,
-            freshness: Freshness::Live,
-            staleness_ms: None,
-        };
-        ToolCallResponse::success(data, meta)
-    }
-
-    #[test]
-    fn returns_cached_exact_when_tier_is_exact() {
-        let resp = resp_with_data(json!({"reused": true, "cache_hit_tier": "exact"}));
-        assert!(matches!(
-            routing_hint_for_payload(&resp),
-            RoutingHint::CachedExact
-        ));
-    }
-
-    #[test]
-    fn returns_cached_warm_when_tier_is_warm() {
-        let resp = resp_with_data(json!({"reused": true, "cache_hit_tier": "warm"}));
-        assert!(matches!(
-            routing_hint_for_payload(&resp),
-            RoutingHint::CachedWarm
-        ));
-    }
-
-    #[test]
-    fn returns_legacy_cached_when_reused_without_tier() {
-        let resp = resp_with_data(json!({"reused": true}));
-        assert!(matches!(
-            routing_hint_for_payload(&resp),
-            RoutingHint::Cached
-        ));
-    }
-
-    #[test]
-    fn returns_async_when_job_id_present() {
-        let resp = resp_with_data(json!({"job_id": "j-1"}));
-        assert!(matches!(
-            routing_hint_for_payload(&resp),
-            RoutingHint::Async
-        ));
-    }
-
-    #[test]
-    fn returns_sync_when_neither_cache_nor_async() {
-        let resp = resp_with_data(json!({"foo": "bar"}));
-        assert!(matches!(routing_hint_for_payload(&resp), RoutingHint::Sync));
     }
 }
